change bash shebang for compatibility & check for dependencies on install
The old Bash shebang doesn't work on any distro that doesn't put Bash in the standard folder, like NixOS or Guix.

It's also hard for end-users to know that they're missing kdialog/wine/pw if the installer doesn't inform them that it's missing. It's just for ease of use for new users.